requirements.txt
python-dotenv
streamlit
google-generativeai
pathlib
youtube_transcript_api